sw, .uno:InsertSection: add a new Content parameter and accept HTML there
When Zotero uses refmarks to represent citations, then a section is used
for the bibliography. While it was possible to insert refmarks with a
given name & content, inserting a section with a provided HTML content
was not possible.

A section (and not a refmark) is needed for the bibliography, as a
refmark is a text attribute, so it must start and end in the same
paragraph.

Fix this by adding a new, optional Content parameter to
.uno:InsertSection, which allows providing multi-paragraph, formatted
text.

This is somewhat similar to what commit
9de8a12765c4384a7764fb35cc7fd85518dc3563 (sw, .uno:InsertBookmark: add a
new BookmarkText parameter and accept HTML there, 2022-12-06) did for
the bookmark case.

Change-Id: I0ccbc8a765fed4e019783d0b743c1cd2d0596093
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145188
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins