Revert "tdf#56258 svx SetOutline: convert to toggle"
This reverts 7.6 commit 8b3a5b5d08f820bd049c7deb269ddcd39d822b6e.

This causes too many unexpected things for the user.
The biggest problem is that it does not join
the previous list like the other numbering does.

I suppose I could fix that by first calling a UNO
command that joins/turns on numbering.

However, that is a bit more dangerous with outline,
since it modifies spacing / numbering on all levels,
and joining can unexpected skip empty paragraphs etc.

So best to just always treat the button as a dropdown
to indicate to the user that they are applying an outline change.

For the present I decided to leave the lighted-up
aspect in place, so show that the outline is in effect.

In terms of removing the list,
which was the main reason for the change in the first place,
there are several workarounds. The user has the option
to go to "more choices" and remove, or select an all-1s choice
and then remove that. Using the bullets button also is
an easy option since it seems less restricted than numbering.

The overall risk is no different from before,
except that it is easier to make those kinds of levels
now that the outline button is always available on the toolbar.

Change-Id: I64e13f203ae65b2d2b28560ba747ca99d1456578
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145240
Tested-by: Justin Luth <jluth@mail.com>
Reviewed-by: Justin Luth <jluth@mail.com>