sw: add a new .uno:UpdateSections command
There was LOK API to insert a new section with provided HTML content and
to query it, but there was no LOK API to update such created section
with new names/contents.

This is needed in case Zotero wants to store citations with refmarks, in
which case it wants to store the bibliography with sections.

Introduce a .uno:UpdateSections UNO command that can do this: the
sections will be renamed if necessary & the content will be updated. The
content update is reasonably straightforward, because the section always
contains at least one empty paragraph, so there is no danger in simply
deleting the old content before inserting the new one.

This is similar to 4382baf4ef9077b4c940edd9fa9a4a0f75d3c173 (sw,
UpdateFields: add new TypeName, NamePrefix and Fields parameters,
2023-01-04), but that was for refmarks / citations, this is for sections
/ bibliography.

Change-Id: Idde6d5ed1b0f25f40df502bb6b7e7ca590ef9151
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145249
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins