ThemeDialog added which allows to change the theme used by the doc.
ThemeDialog is a common dialog that can be used to select the
theme used by the document. Currently it only implements colors
but in the future also the fonts and formats (for shapes) will
be adde.

The IThemeColorChanger interface is used by the dialog to change
the actual color values inside the document. For the writer the
existing ThemeColorChanger is now implementing the interface.

The dialog is accessible in Writer at Format -> Theme... in the
main menu.

Change-Id: I23c7dc9668cdc5427f36d604a76c433d6dbef497
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145264
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>