sw: add a new .uno:DeleteTextFormFields UNO command
Users sometimes want to "unlink" their citations, which means deleting
the fieldmarks (if fieldmarks are used to represent citations, e.g.
with Zotero), which means keeping the field result as-is, but removing
the field metadata and the actual field start/seprator/end characters.

Do this similar to .uno:TextFormFields, which can do an update of such
fieldmarks, i.e. add the ability to filter for a certain type and field
command prefix. This is meant to allow removal af all fieldmark that
belongs to one one feature, e.g. Zotero.

This is similar to 6aaa2fe4637404082a10ba786af3872a805752f8 (sw: add a
new .uno:TextFormFields UNO command, 2022-11-28), but this is about
deleting (the field commands, not the result), while that was about
updating.

The same for bookmarks & refmarks are not yet supported.

Change-Id: I02548b030b1822f7b36d3bc5ff9553d728f065c2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145272
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins