sw lok: expose name of bookmark under cursor
It was possible to get the names of all bookmarks, but you could not get
the name of the (innermost) bookmark under the current cursor.

Getting the name of the current bookmark is useful for Zotero: if we
already have a citation and want to insert one more, then we should turn
the current citation into a citation cluster.

Fix the problem by adding an API similar to what commit
6a7c6c01dd41159ff7c7361d764645ddff98419f (sw lok: expose field type &
command of fieldmark under cursor, 2023-01-05) did, but here we deal
with bookmarks, not fieldmarks.

Handle the actual bookmark lookup in MarkManager, such functionality
looks useful outside LOK as well.

Change-Id: Ic5b9b36fda243c5d7d360fa03745b3e121b67b06
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145323
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins