sw: rename getFieldmarkFor() to getInnerFieldmarkFor()
The same change was made for getBookmarkFor()
in https://gerrit.libreoffice.org/c/core/+/145412

Because otherwise it's quite confusing
that we have a For() and an At()
which could only be differentiated by a code read.

Also improve getInnerFieldmarkFor() a tiny bit,
so we process the first hit only once.

Suggested at
<https://gerrit.libreoffice.org/c/core/+/145348/1#message-286262286f234823b390e8f962e3ba11f5fa71b2>.

Change-Id: I47e815eea0b8ac0df4957ac0d224acb6c5975b8e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145486
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>