tdf#131386 writerfilter: RTF import paragraph mark formatting
One obstacle was remove in commit 6e22db068fa5e855e8df6c0233e076e6e84b98e1

To fix it, when \par is handled, runProps() just has to be called
unconditionally.

Now one test fails with:

rtfexport5.cxx:438:Assertion
Test name: testTdf107413::Load_Verify_Reload_Verify
equality assertion failed
- Expected: 14578
- Actual  : 14698

The spacing above the first paragraph in the table in the footer
is not imported - because this is recorded in a buffer, and when replaying,
getProperties() is called with Id 0, and not LN_Value_ST_StyleType_character
as when it's called immediately without buffering.

Change-Id: If86ed2fa88eb2ac0eafd7a8be411bbfa341619b7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145490
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>