tdf#153009 android: Include more doc layout related XML files
instdir/share/config/soffice.cfg/simpress/objectlist.xml
contains default positions for elements that are used
in new Impress presentations, but that file wasn't
included in the Android app/APK previously, so
couldn't be read from in `SdDrawDocument::InitObjectVector`.
As a result, the title and text were misplaced when creating a
new presentation in Android Viewer.

Include all of the related Impress XML files into the app to
make this work as expected.

(At least a failed attempt to read from `layoutlist.xml`
was also observed while debugging the issue mentioned
above.)

Change-Id: Iba1ac68c0b983d3e465515f07b3fc67162a0ee3f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145621
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>