sw: .uno:TextFormField: handle Endnote as a value for the Wrapper parameter
This is similar to the fieldmark-in-footnote case, but here we need to
make sure that the layout is calculated before the cursor attempts to
jump to the endnote, otherwise the fieldmark will be inserted before the
endnote anchor, not in the endnote body.

The move of StartAction() / EndAction() calls can be done
unconditionally, it's not a problem for the existing footnote case.

Change-Id: I772d27937456978ca6aa01148eff3163d8877208
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145660
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins