ScriptForge - (SF_Calc) sort ranges on more than 3 keys
The Calc.SortRange() method sorts
the given range on any number of columns/rows.
The sorting order may vary by column/row.

The sorting algorithm allows for maximum 3 keys.

When the number of sort keys is > 3 then the range
is sorted several times, by groups of 3 keys,
starting from the last key.

In this context the algorithm used by Calc
to sort ranges is presumed STABLE,
i.e. it maintains the relative order of records
with equal keys.

Change-Id: If7f4920f7ab8f8ffb71edf648ed9accc8eb62dce
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145681
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins