tdf#114537 docx export: export conditional text with " as '
I couldn't find any way  in MS Word to escape a double-quote
that is inside of a conditonal text result.
So just "emulate" that with a single-quote,
which is a common substitute at least in the English world.

Hmm, Google says that { QUOTE 34 } would do the trick,
but it didn't work for me in Word 2010.

Change-Id: Ib9833e482911809c86b4e300533ea0aedbeed71c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145685
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>