tdf#114537 doc import: trim switches before evaluating FIELD_IF
Although this function is only used for ww8 import (and qa tests),
it is documented as being more generic. So I decided to just trim
at the source and not try to introduce any MS-isms into the parse function.
Something similar will be needed for DOCX,
but DOCX import for FIELD_IF is completely missing.

Change-Id: I822b400e3e53abd953f4c382947f0e80ae62b234
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145691
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>