sw XHTML export: use CSS instead of <center> for tables
<center> is not valid XHTML, have to use CSS styling instead.

HTML export uses <center> by default around tables where the alignment
is center.

Fix the problem by avoiding <center> in the XHTML case and set the left
and right margin to auto, which means:

	If the values of margin-left and margin-right are both auto, the
	calculated space is evenly distributed.

according to
<https://developer.mozilla.org/en-US/docs/Web/CSS/margin-left>.

The import will be adjusted to recognize the new markup in a follow-up
change.

Change-Id: I51e3507e9cde713f961b783378d66db59194a6ca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145814
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins